backend/requirements/base.txt
Django
django-cors-headers
django-environ
psycopg2-binary
whitenoise